prompts_crawler_ai/requirements.txt
beautifulsoup4
python-dotenv
google-generativeai
lxml
requests
selenium
undetected-chromedriver
Pillow